refactoring ui component dir
