+++ src/app/layout.tsx
@@ -4,11 +4,11 @@ import { ToastContainer } from "react-toastify";
 import { quicksand } from "@/utils/ui";
 import { CoreProvider } from "@/lib/providers";
 import { getAuthSession } from "@/lib/auth/auth.options";
-import { Box } from "@/tw-styled/ui";
 import { Footer, Header } from "@/shared/components";
 import "./globals.css";
 import "react-toastify/ReactToastify.min.css";
 import "react-tooltip/dist/react-tooltip.css";
+import { Box } from "@/ui/components";
 
 export const metadata: Metadata = {
   title: "Hermes",
