+++ src/features/authentication/molecules/SignInForm.tsx
@@ -1,8 +1,8 @@
 "use client";
-import { Form2 } from "@/tw-styled/ui";
 import { formHooks } from "@/shared/hooks";
 import { FormProvider } from "react-hook-form";
 import { AuthInputs } from "../atoms/Inputs";
+import { Button, Form } from "@/ui/components";
 
 export const SignInForm = () => {
   const { submitHandler, methods } = formHooks.useSignInForm();
@@ -10,16 +10,13 @@ export const SignInForm = () => {
   return (
     <div>
       <FormProvider {...methods}>
-        <Form2
+        <Form
           submitHandler={submitHandler}
-          buttonProps={{
-            text: "Sign In",
-            variant: "primary"
-          }}
-          style={{ formStyles: { backgroundColor: "primary" } }}
+          style={{ form: { backgroundColor: "primary", width: "md", padding: "lg" } }}
         >
           <AuthInputs />
-        </Form2>
+          <Button options={{ variant: "primary" }}>Sign In</Button>
+        </Form>
       </FormProvider>
     </div>
   );
