+++ src/features/authentication/molecules/SwitchForm.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Box } from "@/tw-styled/ui";
+import { Box } from "@/ui/components";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
