+++ src/features/dashboard/super/admin-permissions/templates/AdminPermissions.tsx
@@ -1,8 +1,9 @@
 "use client";
-import { Pulse, Box } from "@/tw-styled/ui";
+import { Pulse } from "@/tw-styled/ui";
 import { PermissionList } from "../organisms";
 import { PermissionCard } from "../molecules";
 import { useFetchPermissions } from "./AdminPermissions.hooks";
+import { Box } from "@/ui/components";
 
 export const AdminPermissions = () => {
   const { data, isLoading } = useFetchPermissions();
