+++ src/features/dashboard/super/authorized-admins/molecules/AdminCard.tsx
@@ -1,8 +1,9 @@
-import { Box, Button, Icon } from "@/tw-styled/ui";
+import { Icon } from "@/tw-styled/ui";
 import { AuthorizedAdmin } from "@prisma/client";
 import { useDeleteAuthorizedAdmin } from "../templates/AuthorizedAdmins.hooks";
 import { useState } from "react";
 import { Modal } from "@/shared/components";
+import { Box, Button } from "@/ui/components";
 
 export const AdminCard = ({ admin }: { admin: AuthorizedAdmin }) => {
   /*
