+++ src/features/dashboard/super/authorized-admins/organisms/AdminList.tsx
@@ -1,8 +1,9 @@
-import { Box, Pulse } from "@/tw-styled/ui";
+import { Pulse } from "@/tw-styled/ui";
 import { useFetchAuthorizedAdmins } from "../templates/AuthorizedAdmins.hooks";
 import { useIcons } from "@/tw-styled/tools";
 import { useTooltip } from "@/shared/hooks";
 import { AdminCard } from "../molecules";
+import { Box } from "@/ui/components";
 
 export const AdminList = () => {
   const { data, isLoading } = useFetchAuthorizedAdmins();
