+++ src/features/dashboard/super/authorized-admins/organisms/AuthorizedAdminForm.tsx
@@ -4,9 +4,10 @@ import {
   useAddAuthorizedAdmin,
   useAuthorizedAdminsForm
 } from "../templates/AuthorizedAdmins.hooks";
-import { Box, FieldError, Form, Input, Label, SubmitButton } from "@/tw-styled/ui";
+import { FieldError, Input, Label, SubmitButton } from "@/tw-styled/ui";
 import { useIcons } from "@/tw-styled/tools";
 import { useTooltip } from "@/shared/hooks";
+import { Box, Form } from "@/ui/components";
 
 export const AuthorizedAdminForm = () => {
   const { mutate } = useAddAuthorizedAdmin();
