+++ src/shared/components/footer/FooterItem.tsx
@@ -1,5 +1,6 @@
 import { Children } from "@/tw-styled/types";
-import { Box, Heading } from "@/tw-styled/ui";
+import { Heading } from "@/tw-styled/ui";
+import { Box } from "@/ui/components";
 
 type FooterItemProps = {
   children?: Children;
