--- src/tw-styled/ui/card/index.ts
@@ -1,1 +1,0 @@
-export * from "./Card";
