--- src/tw-styled/ui/form2/index.ts
@@ -1,1 +1,0 @@
-export * from "./Form";
