+++ src/ui/components/box/Box.types.ts
@@ -1,0 +1,7 @@
+import { BaseStyles, WrapperElements } from "@/tw-styled/types";
+
+export type BoxProps = {
+  children?: React.ReactNode;
+  as?: WrapperElements;
+  style?: BaseStyles;
+};
